+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+- Add `Moar.String.lorem/1` which generates a "lorem ipsum" string of the given length.
+
 ## 1.59.2
 
 - `Moar.UUID.valid?/1` requires the first segment to be 8 hexadecimal characters.